Move DCAP attestation to the attestation-handler
